requirements.txt
apache-airflow==2.5.1
requests==2.26.0
psycopg2-binary==2.9.5
pandas==1.3.3
packaging>=20.0
SQLAlchemy>=1.4.0
python-dotenv==1.0.0
sendgrid==6.9.7
openpyxl